fix: allow running nx run webapp:graphql:download-schema if the contentful env vars are not set

+++ packages/webapp-libs/webapp-contentful/scripts/download-graphql-schema.sh
@@ -2,6 +2,21 @@
 
 export $(egrep -v '^#' .env | xargs)
 
+if [ "$VITE_CONTENTFUL_SPACE" == "<CHANGE_ME>" ]; then
+    echo "VITE_CONTENTFUL_SPACE has not been set. Skipping..."
+    exit 0
+fi
+
+if [ "$VITE_CONTENTFUL_ENV" == "<CHANGE_ME>" ]; then
+    echo "VITE_CONTENTFUL_ENV has not been set. Skipping..."
+    exit 0
+fi
+
+if [ "$VITE_CONTENTFUL_TOKEN" == "<CHANGE_ME>" ]; then
+    echo "VITE_CONTENTFUL_TOKEN has not been set. Skipping..."
+    exit 0
+fi
+
 CONTENTFUL_URL="https://graphql.contentful.com/content/v1/spaces/$VITE_CONTENTFUL_SPACE/environments/$VITE_CONTENTFUL_ENV?access_token=$VITE_CONTENTFUL_TOKEN"
 
 rm  -f ./src/contentful/__generated/types.ts
